+++ plugins/autocomplete/plugin.js
@@ -175,6 +175,15 @@
 		this.throttle = config.throttle !== undefined ? config.throttle : 20;
 
 		/**
+		 * Indicates if a following space should be added after inserted match into an editor.
+		 *
+		 * @since 4.20.0
+		 * @readonly
+		 * @property {Boolean} [followingSpace]
+		 */
+		this.followingSpace = config.followingSpace;
+
+		/**
 		 * The autocomplete view instance.
 		 *
 		 * @readonly
@@ -424,11 +433,20 @@
 			}
 
 			var item = this.model.getItemById( itemId ),
-				editor = this.editor;
+				editor = this.editor,
+				html = this.getHtmlToInsert( item );
+
+			// Insert space after accepting match (#2008).
+			html += this.followingSpace ? '&nbsp;' : '';
 
 			editor.fire( 'saveSnapshot' );
 			editor.getSelection().selectRanges( [ this.model.range ] );
-			editor.insertHtml( this.getHtmlToInsert( item ), 'text' );
+			editor.insertHtml( html, 'text' );
+
+			if ( this.followingSpace ) {
+				removeLeadingSpace( editor );
+			}
+
 			editor.fire( 'saveSnapshot' );
 		},
 
@@ -1517,6 +1535,21 @@
 		}, {} );
 	}
 
+	function removeLeadingSpace( editor ) {
+		var selection = editor.getSelection(),
+			nextNode = selection.getRanges()[ 0 ].getNextNode( function( node ) {
+				return Boolean( node.type == CKEDITOR.NODE_TEXT && node.getText() );
+			} );
+
+		if ( nextNode && nextNode.getText().match( /^\s+/ ) ) {
+			var range = editor.createRange();
+
+			range.setStart( nextNode, 0 );
+			range.setEnd( nextNode, 1 );
+			range.deleteContents();
+		}
+	}
+
 	/**
 	 * Abstract class describing the definition of the [Autocomplete](https://ckeditor.com/cke4/addon/autocomplete) plugin configuration.
 	 *
@@ -1624,6 +1657,12 @@
 	/**
 	 * @inheritdoc CKEDITOR.plugins.autocomplete#outputTemplate
 	 * @property {String} [outputTemplate]
+	 */
+
+	/**
+	 * @inheritdoc CKEDITOR.plugins.autocomplete#followingSpace
+	 * @since 4.20.0
+	 * @property {Boolean} [followingSpace]
 	 */
 
 	/**